requirements.txt
django==4.2.18
djangorestframework==3.15.2
Pillow==11.1.0
drf_yasg==1.21.8
django-rest-swagger==2.2.0

# dev
django-extensions==3.2.3